Relocate ServiceNameRegistry to wildfly-clustering-service module.
